Added support for bincode serialized arguments

+++ example/src/data.rs
@@ -1,12 +1,12 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Default, Deserialize, Serialize)]
+#[derive(Debug, Default, Deserialize, Serialize)]
 pub struct Contact {
-  pub id: u32,
+  pub id: u64,
   pub name: String,
 }
 
-#[derive(Default, Deserialize, Serialize)]
+#[derive(Debug, Default, Deserialize, Serialize)]
 pub struct Error {
   message: String,
 }
